feat(io_uring): implement thread per core io_uring (#2299)
BREAKING. 

This PR implements the thread per core (shared nothing) architecture, using io_uring as I/O backend. 

---------

+++ core/cli/src/args/permissions/mod.rs
@@ -49,7 +49,7 @@ impl From<PermissionsArgs> for Option<Permissions> {
             .stream
             .into_iter()
             .map(|s| (s.stream_id, s.into()))
-            .collect::<AHashMap<u32, StreamPermissions>>();
+            .collect::<AHashMap<usize, StreamPermissions>>();
 
         match (value.global, stream_permissions.is_empty()) {
             (Some(global), true) => Some(Permissions {
