+++ core/server/src/streaming/polling_consumer.rs
@@ -22,23 +22,23 @@ use std::fmt::{Display, Formatter};
 
 #[derive(Debug, PartialEq, Copy, Clone)]
 pub enum PollingConsumer {
-    Consumer(u32, u32),      // Consumer ID + Partition ID
-    ConsumerGroup(u32, u32), // Consumer Group ID + Member ID
+    Consumer(usize, usize),      // Consumer ID + Partition ID
+    ConsumerGroup(usize, usize), // Consumer Group ID + Member ID
 }
 
 impl PollingConsumer {
-    pub fn consumer(consumer_id: &Identifier, partition_id: u32) -> Self {
+    pub fn consumer(consumer_id: &Identifier, partition_id: usize) -> Self {
         PollingConsumer::Consumer(Self::resolve_consumer_id(consumer_id), partition_id)
     }
 
-    pub fn consumer_group(consumer_group_id: u32, member_id: u32) -> Self {
+    pub fn consumer_group(consumer_group_id: usize, member_id: usize) -> Self {
         PollingConsumer::ConsumerGroup(consumer_group_id, member_id)
     }
 
-    pub fn resolve_consumer_id(identifier: &Identifier) -> u32 {
+    pub fn resolve_consumer_id(identifier: &Identifier) -> usize {
         match identifier.kind {
-            IdKind::Numeric => identifier.get_u32_value().unwrap(),
-            IdKind::String => hash::calculate_32(&identifier.value),
+            IdKind::Numeric => identifier.get_u32_value().unwrap() as usize,
+            IdKind::String => hash::calculate_32(&identifier.value) as usize,
         }
     }
 }
@@ -75,7 +75,7 @@ mod tests {
 
         assert_eq!(
             polling_consumer,
-            PollingConsumer::Consumer(consumer_id_value, partition_id)
+            PollingConsumer::Consumer(consumer_id_value as usize, partition_id)
         );
     }
