+++ core/server/src/streaming/users/permissioner.rs
@@ -24,11 +24,11 @@ use iggy_common::{GlobalPermissions, Permissions, StreamPermissions};
 #[derive(Debug, Default)]
 pub struct Permissioner {
     pub(super) users_permissions: AHashMap<UserId, GlobalPermissions>,
-    pub(super) users_streams_permissions: AHashMap<(UserId, u32), StreamPermissions>,
+    pub(super) users_streams_permissions: AHashMap<(UserId, usize), StreamPermissions>,
     pub(super) users_that_can_poll_messages_from_all_streams: AHashSet<UserId>,
     pub(super) users_that_can_send_messages_to_all_streams: AHashSet<UserId>,
-    pub(super) users_that_can_poll_messages_from_specific_streams: AHashSet<(UserId, u32)>,
-    pub(super) users_that_can_send_messages_to_specific_streams: AHashSet<(UserId, u32)>,
+    pub(super) users_that_can_poll_messages_from_specific_streams: AHashSet<(UserId, usize)>,
+    pub(super) users_that_can_send_messages_to_specific_streams: AHashSet<(UserId, usize)>,
 }
 
 impl Permissioner {
