+++ core/server/src/streaming/users/permissioner_rules/consumer_groups.rs
@@ -23,8 +23,8 @@ impl Permissioner {
     pub fn create_consumer_group(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.get_topic(user_id, stream_id, topic_id)
     }
@@ -32,8 +32,8 @@ impl Permissioner {
     pub fn delete_consumer_group(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.get_topic(user_id, stream_id, topic_id)
     }
@@ -41,8 +41,8 @@ impl Permissioner {
     pub fn get_consumer_group(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.get_topic(user_id, stream_id, topic_id)
     }
@@ -50,8 +50,8 @@ impl Permissioner {
     pub fn get_consumer_groups(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.get_topic(user_id, stream_id, topic_id)
     }
@@ -59,8 +59,8 @@ impl Permissioner {
     pub fn join_consumer_group(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.get_topic(user_id, stream_id, topic_id)
     }
@@ -68,8 +68,8 @@ impl Permissioner {
     pub fn leave_consumer_group(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.get_topic(user_id, stream_id, topic_id)
     }
