+++ core/server/src/streaming/users/permissioner_rules/consumer_offsets.rs
@@ -23,8 +23,8 @@ impl Permissioner {
     pub fn get_consumer_offset(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.poll_messages(user_id, stream_id, topic_id)
     }
@@ -32,8 +32,8 @@ impl Permissioner {
     pub fn store_consumer_offset(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.poll_messages(user_id, stream_id, topic_id)
     }
@@ -41,8 +41,8 @@ impl Permissioner {
     pub fn delete_consumer_offset(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.poll_messages(user_id, stream_id, topic_id)
     }
