+++ core/server/src/streaming/users/permissioner_rules/partitions.rs
@@ -23,8 +23,8 @@ impl Permissioner {
     pub fn create_partitions(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.update_topic(user_id, stream_id, topic_id)
     }
@@ -32,8 +32,8 @@ impl Permissioner {
     pub fn delete_partitions(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.update_topic(user_id, stream_id, topic_id)
     }
