+++ core/server/src/streaming/users/permissioner_rules/segments.rs
@@ -22,8 +22,8 @@ impl Permissioner {
     pub fn delete_segments(
         &self,
         user_id: u32,
-        stream_id: u32,
-        topic_id: u32,
+        stream_id: usize,
+        topic_id: usize,
     ) -> Result<(), IggyError> {
         self.update_topic(user_id, stream_id, topic_id)
     }
